pyproject.toml
[project]
name = "qlass"
version = "0.1.3"
description = "VQE on linear optical circuits"
authors = [{name = "Farrokh Labib", email = "farrokh@unitary.foundation"}]
requires-python = ">=3.9,<3.14"

dependencies = [
"qiskit==1.4.2",
"qiskit_aer",
"numpy",
"perceval-quandela==0.13.0",
"openfermion",
"openfermionpyscf",
"pyscf",
"tqdm",
"numba"
]

readme = "README.md"

license = "Apache License"
license-files = ["LICENSE"]

[tool.poetry.dev-dependencies]
pytest = "^8.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[project.urls]
Homepage = "https://github.com/unitaryfoundation/qlass"
Issues = "https://github.com/unitaryfoundation/qlass/issues"
